rm claim page and refactor used modals
