--- apps/main-landing/src/app/claim/constants.ts
@@ -1,374 +1,0 @@
-export const CLAIM_CONTENT = {
-  CHECK_ELIGIBILITY: 'check-eligibility',
-  CLAIM: 'claim',
-  NOT_ELIGIBLE: 'not-eligible',
-  VESTING_PLANS: 'vesting-plans',
-  CLAIM_SUCCESSFUL: 'claim-successful',
-  LETTER: 'letter',
-  ABOUT_IDRISS: 'about-idriss',
-} as const;
-
-export type ClaimPageContent =
-  (typeof CLAIM_CONTENT)[keyof typeof CLAIM_CONTENT];
-
-export const CLAIM_ABI = [
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: '_token',
-        type: 'address',
-      },
-      {
-        internalType: 'address',
-        name: '_owner',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'constructor',
-  },
-  {
-    inputs: [],
-    name: 'AlreadyInitialized',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'NewOwnerIsZeroAddress',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'NoHandoverRequest',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'Unauthorized',
-    type: 'error',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: 'address',
-        name: 'by',
-        type: 'address',
-      },
-      {
-        indexed: false,
-        internalType: 'address',
-        name: 'token',
-        type: 'address',
-      },
-      {
-        indexed: false,
-        internalType: 'address',
-        name: 'to',
-        type: 'address',
-      },
-      {
-        indexed: false,
-        internalType: 'uint256',
-        name: 'total',
-        type: 'uint256',
-      },
-      {
-        indexed: false,
-        internalType: 'uint256[]',
-        name: 'claimIndices',
-        type: 'uint256[]',
-      },
-      {
-        indexed: false,
-        internalType: 'bytes',
-        name: 'memo',
-        type: 'bytes',
-      },
-      {
-        indexed: false,
-        internalType: 'bool',
-        name: 'bonus',
-        type: 'bool',
-      },
-    ],
-    name: 'Claimed',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'pendingOwner',
-        type: 'address',
-      },
-    ],
-    name: 'OwnershipHandoverCanceled',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'pendingOwner',
-        type: 'address',
-      },
-    ],
-    name: 'OwnershipHandoverRequested',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'oldOwner',
-        type: 'address',
-      },
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'newOwner',
-        type: 'address',
-      },
-    ],
-    name: 'OwnershipTransferred',
-    type: 'event',
-  },
-  {
-    inputs: [],
-    name: 'cancelOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'to',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'total',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256[]',
-        name: 'claimIndices',
-        type: 'uint256[]',
-      },
-      {
-        internalType: 'bytes',
-        name: 'signature',
-        type: 'bytes',
-      },
-      {
-        internalType: 'uint256',
-        name: 'expiry',
-        type: 'uint256',
-      },
-      {
-        internalType: 'bytes',
-        name: 'memo',
-        type: 'bytes',
-      },
-    ],
-    name: 'claim',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'claimSigner',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'to',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'total',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256[]',
-        name: 'claimIndices',
-        type: 'uint256[]',
-      },
-      {
-        internalType: 'bytes',
-        name: 'signature',
-        type: 'bytes',
-      },
-      {
-        internalType: 'uint256',
-        name: 'expiry',
-        type: 'uint256',
-      },
-      {
-        internalType: 'bytes',
-        name: 'memo',
-        type: 'bytes',
-      },
-    ],
-    name: 'claimWithBonus',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'pendingOwner',
-        type: 'address',
-      },
-    ],
-    name: 'completeOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'uint256',
-        name: 'claimIndex',
-        type: 'uint256',
-      },
-    ],
-    name: 'isClaimed',
-    outputs: [
-      {
-        internalType: 'bool',
-        name: '',
-        type: 'bool',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'owner',
-    outputs: [
-      {
-        internalType: 'address',
-        name: 'result',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'pendingOwner',
-        type: 'address',
-      },
-    ],
-    name: 'ownershipHandoverExpiresAt',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'result',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'renounceOwnership',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'requestOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'newClaimSigner',
-        type: 'address',
-      },
-    ],
-    name: 'setClaimSigner',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'token',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'newOwner',
-        type: 'address',
-      },
-    ],
-    name: 'transferOwnership',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-];
-
-export const CLAIMER_ADDRESS = '0xfD76c3a2A4534D815c9C5127119e6ea738A28837';
-export const claimSteps = [
-  {
-    step: '01',
-    title: 'Letter',
-  },
-  {
-    step: '02',
-    title: 'About IDRISS',
-  },
-  {
-    step: '03',
-    title: 'Claim your $IDRISS',
-  },
-];
