+++ packages/ui/package.json
@@ -24,6 +24,7 @@
     "./multiselect": "./src/components/multiselect/index.ts",
     "./scroll-area": "./src/components/scroll-area/index.ts",
     "./wallet-connect-modal": "./src/components/wallet-connect-modal/index.ts",
+    "./tx-loading-modal": "./src/components/tx-loading-modal/index.ts",
     "./progress-bar": "./src/components/progress-bar/index.ts",
     "./providers/with-portal": "./src/providers/with-portal.tsx",
     "./utils": "./src/utils/index.ts",
