+++ packages/ui/src/components/tx-loading-modal/index.ts
@@ -1,0 +1,1 @@
+export { TxLoadingModal } from './tx-loading-modal';
