+++ packages/utils/src/index.ts
@@ -9,6 +9,7 @@ export {
   getChainById,
   isUnrecognizedChainError,
   formatBigNumber,
+  formatNumber,
   getDefaultBlockExplorerUrl,
   getTransactionUrls,
   getSafeNumber,
